enum -> enum class in lexer
This makes enum elements scoped which is actually quite useful as I
prefer the namespacing that enum's give in C++.

+++ asm/main.cpp
@@ -100,7 +100,7 @@ int main(int argc, const char *argv[])
     }
     std::cerr << file_name << ":" << line << ":" << col << ":"
               << lerr_as_cstr(lerr) << std::endl;
-    ret = 255 - lerr;
+    ret = 255 - static_cast<int>(lerr);
     goto end;
   }
   else
